Overall improvements in interop with Windows. Added surface interop, moved from AVX ICD to here to match uniformity and coherence.
