--- src/Shell_W32.c
@@ -1,40 +1,0 @@
-/*
- *          ::::::::  :::       :::     :::     :::::::::  :::::::::   ::::::::
- *         :+:    :+: :+:       :+:   :+: :+:   :+:    :+: :+:    :+: :+:    :+:
- *         +:+    +:+ +:+       +:+  +:+   +:+  +:+    +:+ +:+    +:+ +:+    +:+
- *         +#+    +:+ +#+  +:+  +#+ +#++:++#++: +#+    +:+ +#++:++#:  +#+    +:+
- *         +#+  # +#+ +#+ +#+#+ +#+ +#+     +#+ +#+    +#+ +#+    +#+ +#+    +#+
- *         #+#   +#+   #+#+# #+#+#  #+#     #+# #+#    #+# #+#    #+# #+#    #+#
- *          ###### ###  ###   ###   ###     ### #########  ###    ###  ########
- *
- *                  Q W A D R O   E X E C U T I O N   E C O S Y S T E M
- *
- *                                   Public Test Build
- *                               (c) 2017 SIGMA FEDERATION
- *                             <https://sigmaco.org/qwadro/>
- */
-
-#define _AFX_SYSTEM_C
-#define _AUX_UX_C
-#define _AUX_SHELL_C
-#define _AUX_SHELL_IMPL
-#define _AUX_WINDOW_C
-#define _AUX_WINDOW_IMPL
-//#include "qwadro/../_luna/luna_vm.h"
-//#include "qwadro/../_luna/luna.h"
-#include "AuxOverWin32.h"
-
-_QOW afxError afxIcdHook(afxModule icd, afxUri const* manifest)
-{
-    afxError err = AFX_ERR_NONE;
-    AFX_ASSERT_OBJECTS(afxFcc_MDLE, 1, &icd);
-
-    afxClassConfig sesClsCfg = _AUX_SES_CLASS_CONFIG;
-    sesClsCfg.fixedSiz = sizeof(AFX_OBJ(afxSession));
-    sesClsCfg.ctor = (void*)_QowSesCtorCb;
-    sesClsCfg.dtor = (void*)_QowSesDtorCb;
-
-    _AuxImplementShell(icd, &sesClsCfg);
-
-    return err;
-}
