--- src/Widget_W32.c
@@ -1,47 +1,0 @@
-/*
- *          ::::::::  :::       :::     :::     :::::::::  :::::::::   ::::::::
- *         :+:    :+: :+:       :+:   :+: :+:   :+:    :+: :+:    :+: :+:    :+:
- *         +:+    +:+ +:+       +:+  +:+   +:+  +:+    +:+ +:+    +:+ +:+    +:+
- *         +#+    +:+ +#+  +:+  +#+ +#++:++#++: +#+    +:+ +#++:++#:  +#+    +:+
- *         +#+  # +#+ +#+ +#+#+ +#+ +#+     +#+ +#+    +#+ +#+    +#+ +#+    +#+
- *         #+#   +#+   #+#+# #+#+#  #+#     #+# #+#    #+# #+#    #+# #+#    #+#
- *          ###### ###  ###   ###   ###     ### #########  ###    ###  ########
- *
- *                  Q W A D R O   E X E C U T I O N   E C O S Y S T E M
- *
- *                                   Public Test Build
- *                               (c) 2017 SIGMA FEDERATION
- *                             <https://sigmaco.org/qwadro/>
- */
-
-#include "AuxOverWin32.h"
-
-_QOW afxError _QowWidDtorCb(afxWidget wid)
-{
-    afxError err = AFX_ERR_NONE;
-    AFX_ASSERT_OBJECTS(afxFcc_WID, 1, &wid);
-
-    AFX_ASSERT(AfxGetTid() == AfxGetObjectTid(wid));
-
-    _AUX_WID_CLASS_CONFIG.dtor(wid);
-
-    return err;
-}
-
-_QOW afxError _QowWidCtorCb(afxWidget wid, void** args, afxUnit invokeNo)
-{
-    afxError err = AFX_ERR_NONE;
-    AFX_ASSERT_OBJECTS(afxFcc_WID, 1, &wid);
-
-    afxWindow wnd = args[0];
-    AFX_ASSERT_OBJECTS(afxFcc_WND, 1, &wnd);
-    afxWidgetConfig const* cfg = AFX_CAST(afxWidgetConfig const*, args[1]) + invokeNo;
-    AFX_ASSERT(cfg);
-    
-    if (_AUX_WID_CLASS_CONFIG.ctor(wid, args, invokeNo)) AfxThrowError();
-    else
-    {
-        
-    }
-    return err;
-}
